refactor: Organise EC2 app pattern files
Rename `src/patterns/ec2-app.ts` to `src/patterns/ec2-app/base.ts`.
