+++ src/patterns/ec2-app/base.ts
@@ -5,30 +5,30 @@ import { Port } from "@aws-cdk/aws-ec2";
 import { ApplicationProtocol } from "@aws-cdk/aws-elasticloadbalancingv2";
 import { Bucket } from "@aws-cdk/aws-s3";
 import { Duration, Tags, Token } from "@aws-cdk/core";
-import { Stage } from "../constants";
-import { AccessScope } from "../constants/access";
-import { SSM_PARAMETER_PATHS } from "../constants/ssm-parameter-paths";
-import { TagKeys } from "../constants/tag-keys";
-import { GuCertificate } from "../constructs/acm";
-import type { GuUserDataProps } from "../constructs/autoscaling";
-import { GuAutoScalingGroup, GuUserData } from "../constructs/autoscaling";
-import type { Http5xxAlarmProps, NoMonitoring } from "../constructs/cloudwatch";
-import { Gu5xxPercentageAlarm, GuUnhealthyInstancesAlarm } from "../constructs/cloudwatch";
-import type { GuStack } from "../constructs/core";
-import { GuStringParameter } from "../constructs/core";
-import { AppIdentity } from "../constructs/core/identity";
-import { GuSecurityGroup, GuVpc, SubnetType } from "../constructs/ec2";
-import type { GuInstanceRoleProps } from "../constructs/iam";
-import { GuGetPrivateConfigPolicy, GuInstanceRole } from "../constructs/iam";
+import { Stage } from "../../constants";
+import { AccessScope } from "../../constants/access";
+import { SSM_PARAMETER_PATHS } from "../../constants/ssm-parameter-paths";
+import { TagKeys } from "../../constants/tag-keys";
+import { GuCertificate } from "../../constructs/acm";
+import type { GuUserDataProps } from "../../constructs/autoscaling";
+import { GuAutoScalingGroup, GuUserData } from "../../constructs/autoscaling";
+import type { Http5xxAlarmProps, NoMonitoring } from "../../constructs/cloudwatch";
+import { Gu5xxPercentageAlarm, GuUnhealthyInstancesAlarm } from "../../constructs/cloudwatch";
+import type { GuStack } from "../../constructs/core";
+import { GuStringParameter } from "../../constructs/core";
+import { AppIdentity } from "../../constructs/core/identity";
+import { GuSecurityGroup, GuVpc, SubnetType } from "../../constructs/ec2";
+import type { GuInstanceRoleProps } from "../../constructs/iam";
+import { GuGetPrivateConfigPolicy, GuInstanceRole } from "../../constructs/iam";
 import {
   GuApplicationLoadBalancer,
   GuApplicationTargetGroup,
   GuHttpsApplicationListener,
-} from "../constructs/loadbalancing";
-import type { AppAccess, InternalAccess, RestrictedAccess } from "../types/access";
-import type { GuAsgCapacity } from "../types/asg";
-import type { GuDomainName } from "../types/domain-names";
-import { StageAwareValue } from "../types/stage";
+} from "../../constructs/loadbalancing";
+import type { AppAccess, InternalAccess, RestrictedAccess } from "../../types/access";
+import type { GuAsgCapacity } from "../../types/asg";
+import type { GuDomainName } from "../../types/domain-names";
+import { StageAwareValue } from "../../types/stage";
 
 export interface AccessLoggingProps {
   enabled: boolean;
